backend/requirements.txt
slack-bolt>=1.21.3
slack-sdk>=3.27.2
anthropic>=0.35.0
requests>=2.32.3
pygithub>=2.3.0
python-dotenv>=1.0.1
jira>=3.6.0
pydantic>=2.7.3
pydantic-settings>=2.3.4
structlog>=24.1.0
httpx>=0.27.0
